Migrate to bzlmod and add nix flake for building
- Add flake.nix with FHS-wrapped Bazel for NixOS compatibility
- Migrate from WORKSPACE to MODULE.bazel (bzlmod)
- Update to latest Bazel rules:
  - rules_kotlin 2.3.0
  - rules_jvm_external 6.10
  - rules_java 9.3.0
- Update dependencies:
  - JavaFX 11 -> 21
  - Exposed 0.18.1 (JitPack) -> 0.37.3 (Maven Central)
- Fix Exposed API compatibility in DatabaseRecorder.kt
- Add .bazelrc for C++17 and remote JDK configuration
- Update BUILD files to use @rules_kotlin instead of @io_bazel_rules_kotlin

Build with: nix develop --command bazel build //:Gui
Run with: nix develop --command bazel run //:Gui

+++ src/me/danielschaefer/sensorwrangler/sensors/AntPowerSensor.kt
@@ -95,7 +95,7 @@ class AntPowerSensor : AntPlusSensor<PowerChannel>() {
                 dataPage.riderPosition
             }
             is PowerDataPage10PowerOnly -> {
-                powerMeasurement.addDataPoint(dataPage.instantaneousPower.toDouble())
+                powerMeasurement.addDataPoint(dataPage.instantaneousPower)
                 cadenceMeasurement.addDataPoint(dataPage.instantaneousCadence.toDouble())
 
                 dataPage.pedalPowerDistribution?.let {
